+++ src/models/MiniReel.js
@@ -28,6 +28,7 @@ import DisplayAdCard from './DisplayAdCard.js';
 import RecapCard from './RecapCard.js';
 import PrerollCard from './PrerollCard.js';
 import SlideshowBobCard from './SlideshowBobCard.js';
+import VineVideoCard from './VineVideoCard.js';
 
 const CARD_WHITELIST = ['text', 'video', 'article', 'image', 'displayAd', 'slideshow-bob', 'recap'];
 
@@ -41,6 +42,7 @@ function getCardType(card) {
     case 'rumble':
     case 'embedded':
     case 'adUnit':
+    case 'vine':
         return 'video';
     default:
         return card.type;
@@ -74,6 +76,8 @@ function initialize(whitelist, { experience, standalone, profile }) {
             return new DisplayAdCard(card, experience, profile);
         case 'slideshow-bob':
             return new SlideshowBobCard(card, experience, profile);
+        case 'vine':
+            return new VineVideoCard(card, experience, profile);
         default:
             return new VideoCard(card, experience, profile);
         }